chore: migrate to CDK CLI stage 2 - Layers work
