+++ tests/e2e/lambda_layer/Dockerfile
