+++ tests/e2e/tracer/conftest.py
@@ -1,16 +1,13 @@
-import json
-import os
 from pathlib import Path
 
 import pytest
 
 from tests.e2e.tracer.infrastructure import TracerStack
 
-PWD = Path(__file__).parent
 
-# def infrastructure(request: pytest.FixtureRequest, lambda_layer_arn: str):
 @pytest.fixture(autouse=True, scope="module")
-def infrastructure(request: pytest.FixtureRequest):
+def infrastructure(request: pytest.FixtureRequest, lambda_layer_arn: str):
+    # # def infrastructure(request: pytest.FixtureRequest):
     """Setup and teardown logic for E2E test infrastructure
 
     Parameters
@@ -25,9 +22,7 @@ def infrastructure(request: pytest.FixtureRequest):
     Dict[str, str]
         CloudFormation Outputs from deployed infrastructure
     """
-    # NOTE: Commented out for faster debug as we don't need a Layer yet
-    # stack = TracerStack(handlers_dir=Path(f"{request.path.parent}/handlers"), layer_arn=lambda_layer_arn)
-    stack = TracerStack(handlers_dir=Path(f"{request.path.parent}/handlers"), layer_arn="")
+    stack = TracerStack(handlers_dir=Path(f"{request.path.parent}/handlers"), layer_arn=lambda_layer_arn)
     try:
         yield stack.deploy()
     finally:
